feat: ListingDetails component added
