+++ src/components/ListingDetails/listingDetaild.types.d.ts
@@ -1,0 +1,12 @@
+import { IListingProps } from '../../sections/Listings/listings.types';
+
+/**
+ * Types for ListingDetails
+ * @param {id} string - Id of the listing
+ * @param {data} IListingProps[] - Listing data
+ */
+
+export interface IListingDetails {
+  id: string | undefined;
+  data: IListingProps[] | null;
+}
