+++ src/components/index.ts
@@ -6,3 +6,4 @@ export * from './OAuthComponent';
 export * from './MenuItem';
 export * from './InputComponent';
 export * from './BookingCard';
+export * from './ListingDetails';
